requirements.txt
flask==3.0.2
flask-mail==0.9.1
flask-wtf==1.2.1
email-validator==2.1.1
python-dotenv==1.0.1